A working version that builds against my global .cabal file, contents of which are in 'pkg-list'.  Using:
- ghc 7.4.2 from apt-get
- cabal 1.22.0.0 from 'cabal install cabal-install' via apt-get.

Curiously, it won't build in a sandbox: cabal can't resolve the
dependencies.

This implementation has been refactored to use only

1. Text.CSV (csv-0.1.2)
2. Text.XML.Light (xml-xml-1.3.13)

For this application these are completely adequate.  In fact, the XML
implementation is now much simpler as Text.XML.Light has a
straightforward tree-traversal model that is easier to grok than HXT's
arrows.

The CSV is somewhat less convenient than Cassava as there is no way to
access a row's fields by name, but the workaround was easy (translate
names to column numbers once).

The pivot table code is much cleaner now as it works on a list of rows
rather than the native representation, so should work with any table
(but this has not been tested with XML yet).